랭킹 시스템 수정

+++ src/main/kotlin/org/hobro/musicranker/model/MusicDTO.kt
@@ -8,7 +8,7 @@ data class MusicDTO(
     val videoId: String,
     val title: String,
     val singer: String,
-    val rank: Long?,
+    val rank: Int?,
     val wave: WaveType?,
     val likeCount: Long,
     val dislikeCount: Long
